requirements.txt
transformers 
sentence-transformers 
faiss-cpu